fix: changes 'req' to 'token' in comment to avoid confusion
The comment it about how the 'req' property will be passed to the callback if you set `passReqToCallback` to `true`. The comment said that you'd get `req, token` even if it was false. I've updated it so that it says `token, done`, since you will not receive `req` if it is set to false.